% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_results.R
\name{aggregate_mean_results}
\alias{aggregate_mean_results}
\title{Aggregate and save coverage results from different runs.
Aggregate averaged metric values from different runs.}
\usage{
aggregate_mean_results(metric, results_dirs, digits = 3)
}
\description{
Aggregate and save coverage results from different runs.
Aggregate averaged metric values from different runs.
}
