% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_results.R
\name{aggregate_raw_results}
\alias{aggregate_raw_results}
\title{Aggregate averaged metric values from different runs.}
\usage{
aggregate_raw_results(files)
}
\description{
Aggregate averaged metric values from different runs.
}
