% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_analysis.R
\name{bound_pi}
\alias{bound_pi}
\title{Bound estimated propensity scores away from 0 and 1.}
\usage{
bound_pi(pi, lower = 0.025, upper = 0.975)
}
\arguments{
\item{pi}{Vector of estimated propensity scores in [0, 1].}

\item{lower}{Lower bound.}

\item{upper}{Upper bound.}
}
\value{
The scores, bounded in [lower, upper].
}
\description{
Bound estimated propensity scores away from 0 and 1.
}
