% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_results.R
\name{calc_means}
\alias{calc_means}
\title{Get mean of metric values, for new code.}
\usage{
calc_means(results_dir, metric, write = TRUE)

calc_means(results_dir, metric, write = TRUE)
}
\arguments{
\item{results_dir}{Directory containing a file results.txt which is created
by npDR().}

\item{metric}{One of "bias", "rmse" (which is actually mse), "width", or
"cov" (for coverage).}

\item{write}{Logical. If TRUE, then write output to results_dir.}

\item{var}{One of "bias", "mse", "width", or "cov" (for coverage).}

\item{results_dir}{Directory containing a file results.txt which is created
by npDR().}

\item{write}{Logical. If TRUE, then write output to results_dir.}
}
\value{
A vector of mean values for the chosen var. Means are taken over
all the simulations represented in results.txt.

A vector of mean values for the chosen metric. Means are taken over
all the simulations represented in results.txt.
}
\description{
Get mean of metric values, for new code.

Get mean of metric values, for the old code.
}
