% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_analysis.R
\name{calc_metrics}
\alias{calc_metrics}
\title{Calculate performance metrics for an estimate of the ATE.}
\usage{
calc_metrics(est, effect_size)
}
\arguments{
\item{est}{A named vector with elements "est", the estimated ATE, and "se",
the estimated standard error.}

\item{effect_size}{The true ATE, a number.}
}
\value{
A named vector with elements "bias," "mse," "cov" (coverage), and
"width". The confidence interval is computed as est +/- 1.96*se. The coverage
is 0 or 1, indicating whether the true ATE falls in the interval.
}
\description{
The metrics calculated are the bias, squared error, and the coverage and
width of a 95% confidence interval.
}
