% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_analysis.R
\name{est_aipw}
\alias{est_aipw}
\title{Compute AIPW estimate of ATE.}
\usage{
est_aipw(Y, A, pihat, muhat)
}
\arguments{
\item{Y}{A vector of continuous outcomes of length n.}

\item{A}{A vector of treatment values, in {0, 1}.}

\item{pihat}{A vector of estimated propensity scores. Estimation can be poor
when values are close to 0 or 1, so it may be advisable to bound these in,
say, [0.025, 0.975].}

\item{muhat}{2-column data frame of estimated regression values, where
column 1 = E[Y|A=0, X] and column 2 = E[Y|A=1, X]. The columns must be named
mu0 and mu1.}
}
\value{
A named vector with elements "est", the estimated ATE, and "se", the
standard error, estimated analytically.
}
\description{
Compute AIPW estimate of ATE.
}
