% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_analysis.R
\name{est_gcomp}
\alias{est_gcomp}
\title{Compute regression, aka g-computation, estimate of the ATE.}
\usage{
est_gcomp(muhat)
}
\arguments{
\item{muhat}{2-column data frame of estimated regression values, where
column 1 = E[Y|A=0, X] and column 2 = E[Y|A=1, X]. The columns must be named
mu0 and mu1.}
}
\value{
A named vector with elements "est": the estimated ATE using the
g-computation estimator; and "se", which is always NA and which is a
placeholder for the bootstrap-estimated standard error. This element is
included to make the output of this function equivalent to the output from
other estimator functions.
}
\description{
Compute regression, aka g-computation, estimate of the ATE.
}
