% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_analysis.R
\name{est_gcomp_nonparametric}
\alias{est_gcomp_nonparametric}
\title{Compute g-computation estimate of ATE with nonparametric outcome model.}
\usage{
est_gcomp_nonparametric(
  dat,
  rows = NULL,
  sl_lib,
  covs = "X",
  treat = "A",
  outcome = "Y",
  nsplits = NULL,
  split_inds = NULL,
  ...
)
}
\arguments{
\item{dat}{Data frame containing covariates, treatments, and outcomes.}

\item{rows}{Which rows of the data to include. If NULL, then all rows are
included. This argument is required in order to pass the function to boot().}

\item{sl_lib}{A character vector of learner functions created via
SuperLearner, as in sl_lib <- SuperLearner::create.Learner(...)$names, where
... are the learner type and learner parameters. These are passed to
SuperLearner() to estimate the propensity score function.}

\item{covs}{String representing the names of the covariate columns, which
should be numbered starting with 1. For example, the default "X" means that
the covariate columns are named X1, X2, etc.}

\item{treat}{The name of the treatment column.}

\item{outcome}{The name of the outcome column.}

\item{nsplits}{The number of splits to use for sample splitting, if fold
indices aren't provided to the split_inds argument.}

\item{split_inds}{Fold indices for sample splitting, if nsplits is NULL. One
of split_inds or nsplits must be provided. split_inds takes precedence if
it's not NULL.}

\item{...}{Additional arguments passed to SuperLearner.}
}
\value{
A named vector with elements "est": the estimated ATE using the
g-computation estimator; and "se", which is always NA and which is a
placeholder for the bootstrap-estimated standard error. This element is
included to make the output of this function equivalent to the output from
other estimator functions.
}
\description{
This function is specifically designed to be passed to boot::boot() when
estimating the standard error via bootstrapping.
}
