% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_analysis.R
\name{estimate_gcomp}
\alias{estimate_gcomp}
\title{Compute regression, aka g-computation, estimate of ATE.}
\usage{
estimate_gcomp(muhat)
}
\arguments{
\item{muhat}{2-column data frame of estimated regression values, where
column 1 = \eqn{\hat{E}[Y|A=0, X]} and column 2 = \eqn{\hat{E}[Y|A=1, X]}. The
columns must be named \code{mu0} and \code{mu1}.}
}
\value{
Estimated ATE, the mean of \code{mu1} - \code{mu0}.
}
\description{
Compute regression, aka g-computation, estimate of ATE.
}
