% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_analysis.R
\name{estimate_gcomp_nonpar}
\alias{estimate_gcomp_nonpar}
\title{Compute g-computation estimate of ATE with nonparametric outcome model.}
\usage{
estimate_gcomp_nonpar(dat, rows = NULL, sl, treat = "A",
  outcome = "Y", ...)
}
\arguments{
\item{dat}{Data frame containing covariates, treatments, and outcomes.}

\item{rows}{Which rows of the data to include. If NULL, then all rows are
included. This argument is required in order to pass the function to boot().}

\item{treat}{The name of the treatment column.}

\item{outcome}{The name of the outcome column.}

\item{SL.library}{Vector of names of learners to use when estimating the
outcome model, passed to SuperLearner().}

\item{cv_folds}{Number of cross-validation folds to use, passed to the V
argument of the cvControl list in SuperLearner().}
}
\value{
A named vector with elements "est": the estimated ATE using the
g-computation estimator; and "se", which is always NA and which is a
placeholder for the bootstrap-estimated standard error. This element is
included to make the output of this function equivalent to the output from
other estimator functions. The se must be estimated by bootstrapping, using
this function, and then computing the empirical standard error over the
bootstrap samples.
}
\description{
This function is specifically designed to be passed to boot::boot() when
estimating the standard error via bootstrapping.
}
