% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_analysis.R
\name{estimate_gcomp_par}
\alias{estimate_gcomp_par}
\title{Compute regression, aka g-computation, estimate, via linear regression.}
\usage{
estimate_gcomp_par(dat, rows = NULL, covs = "X", treat = "A",
  outcome = "Y")
}
\arguments{
\item{dat}{Data frame containing covariates, treatments, and outcomes.}

\item{rows}{Which rows of the data to include. If NULL, then all rows are
included. This argument is required in order to pass the function to
\code{boot()}.}

\item{covs}{String representing the names of the covariate columns, which
should be numbered starting with 1. For example, the default "X" means that
the covariate columns are named \code{X1}, \code{X2}, etc.}

\item{treat}{The name of the treatment column.}

\item{outcome}{The name of the outcome column.}
}
\value{
A named vector with elements "est": the estimated ATE using the
g-computation estimator; and "se", which is always NA and which is a
placeholder for the bootstrap-estimated standard error. This element is
included to make the output of this function equivalent to the output from
other estimator functions. The se must be estimated by bootstrapping, using
this function, and then computing the empirical standard error over the
bootstrap samples.
}
\description{
This function is specifically designed to be passed to \code{boot::boot()}
when estimating the standard error via bootstrapping.
}
