% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_analysis.R
\name{estimate_ipw}
\alias{estimate_ipw}
\title{Compute IPW estimate of ATE.}
\usage{
estimate_ipw(Y, A, pihat)
}
\arguments{
\item{Y}{A vector of continuous outcomes of length n.}

\item{A}{A vector of treatment values, in {0, 1}.}

\item{pihat}{A vector of estimated propensity scores. Estimation can be poor
when values are close to 0 or 1, so it may be advisable to bound these in,
say, [0.025, 0.975].}
}
\value{
A named vector with elements "est", the estimated ATE, and "se", the
standard error, estimated analytically.
}
\description{
Compute IPW estimate of ATE.
}
