% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_analysis.R
\name{estimate_mu_par}
\alias{estimate_mu_par}
\title{Estimate outcome regression function parametrically, via linear regression.}
\usage{
estimate_mu_par(Y, A, X)
}
\arguments{
\item{Y}{A vector of continuous outcomes of length n.}

\item{A}{A vector of treatment values of length n, in {0, 1}.}

\item{X}{An n-by-p matrix or data frame of covariate values.}
}
\value{
An n-by-2 data frame of estimated regression values E[Y|A=a, X], for
a = 0 (column 1) and a = 1 (column 2).
}
\description{
Estimate outcome regression function parametrically, via linear regression.
}
