% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_analysis.R
\name{estimate_pi_nonpar}
\alias{estimate_pi_nonpar}
\title{Estimate propensity scores nonparametrically using a superlearner.}
\usage{
estimate_pi_nonpar(A, X, sl, ...)
}
\arguments{
\item{A}{A vector of treatment values, in {0, 1}.}

\item{X}{A matrix or data frame of covariate values.}

\item{sl}{A superlearner, an object of class Lrnr_sl from the \code{sl3}
library.}

\item{...}{Additional arguments passed to SuperLearner.}

\item{split}{Logical, whether to perform sample splitting. If TRUE, then the
sample is randomly split in 2; learning is performed one part, and
propensities are estimated on the other. The process is repeated with the
samples switched so that a full vector of estimated propensities is
generated. If FALSE, then learning and estimation are both performed on the
entire sample.}
}
\value{
A vector of estimated propensities in [0, 1].
}
\description{
Estimate propensity scores nonparametrically using a superlearner.
}
