% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_analysis.R
\name{estimate_pi_par}
\alias{estimate_pi_par}
\title{Estimate propensity scores parametrically, via logistic regression.}
\usage{
estimate_pi_par(A, X)
}
\arguments{
\item{A}{A vector of treatment values, in {0, 1}.}

\item{X}{A matrix or data frame of covariate values.}
}
\value{
A vector of estimated treatment propensities P(A = 1|X) in [0, 1].
}
\description{
Estimate propensity scores parametrically, via logistic regression.
}
