% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_analysis.R
\name{generate_folds}
\alias{generate_folds}
\title{Generate indices for CV folds.}
\usage{
generate_folds(n, cv_folds, probs = NULL)
}
\arguments{
\item{n}{Number of rows in the dataset.}

\item{cv_folds}{Number of folds to create.}

\item{probs}{Probabilities associated with each fold. If NULL, then data
points are assigned to each fold with equal probability. Fold memberships are
iid sampled for each data point.}
}
\value{
A list of row indices corresponding to each fold.
}
\description{
Generate indices for CV folds.
}
