% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_results.R
\name{get_dirs}
\alias{get_dirs}
\title{Get directories containing results, for downstream processing.}
\usage{
get_dirs(topdir = ".", exclude = NULL)
}
\arguments{
\item{topdir}{The top directory to look in.}

\item{exclude}{A string representing directories to exclude. If NULL, then
all directories starting with "results" are included. If non-NULL, then all
directories containing this string are excluded.}
}
\value{
A vector of paths to subdirectories in the top directory which start
with "results".
}
\description{
Get directories containing results, for downstream processing.
}
