% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_results.R
\name{metric_means}
\alias{metric_means}
\title{Process the results file that is generated by simulate_inference().}
\usage{
metric_means(results_file, append, vars = c("est", "se", "bias", "mse",
  "width", "cov"))
}
\arguments{
\item{results_dir}{Path to a directory containing a file results.txt
generated by npDR().}

\item{metrics}{Some subset of c("bias", "mse", "width", "cov").}
}
\value{
None. Writes one file per chosen metric to the results_dir.
}
\description{
Computes the mean metric value across a given set of simulations.
}
