% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_analysis.R
\name{estimate_mu_nonpar}
\alias{estimate_mu_nonpar}
\title{Estimate outcome regression function nonparametrically using SuperLearner.}
\usage{
estimate_mu_nonpar(Y, A, X, sl, ...)
}
\arguments{
\item{Y}{A vector of continuous outcomes of length n.}

\item{A}{A vector of treatment values, in {0, 1}.}

\item{X}{A matrix or data frame of covariate values.}

\item{...}{Additional arguments passed to \code{sl3::make_sl3_Task}.}

\item{sl_lib}{A character vector of learner functions created via
SuperLearner, as in sl_lib <- SuperLearner::create.Learner(...)$names, where
... are the learning type and learner parameters. These are passed to
SuperLearner() to estimate the propensity score function.}
}
\value{
An n-by-3 data frame of estimated regression values, where column
1 = \eqn{E[Y|A, X]}, column 2 = \eqn{E[Y|A=0, X]}, and column 3 =
\eqn{E[Y|A=1, X]}.
}
\description{
Outcomes are treated as continuous, meaning that the error distribution in
SuperLearner() is set to "gaussian."
}
