% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_analysis.R
\name{generate_covariates}
\alias{generate_covariates}
\title{Generate covariates X and transform it to induce model misspecification.}
\usage{
generate_covariates(size)
}
\arguments{
\item{size}{Number of samples.}
}
\value{
A list containing two n-by-4 matrices: X, the baseline covariates, and
Z, the transformed version of X.
}
\description{
Data is 4-dimensional multivariate normal with mean 0 and diagonal
covariance.
}
