% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_analysis.R
\name{generate_folds}
\alias{generate_folds}
\title{Generate indices for cross validates folds.}
\usage{
generate_folds(n, cv_folds)
}
\arguments{
\item{n}{Number of rows.}

\item{cv_folds}{Number of folds.}
}
\value{
Vector of indices indicating fold membership.
}
\description{
Generate indices for cross validates folds.
}
