% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_analysis.R
\name{generate_outcomes}
\alias{generate_outcomes}
\title{Generate treatments and outcomes from covariates.}
\usage{
generate_outcomes(X, beta_a, beta_y, effect_size = 6)
}
\arguments{
\item{X}{n-by-p matrix of covariate data.}

\item{beta_a}{Vector of logistic coefficients of length p, relating the
propensity scores to X.}

\item{beta_y}{Vector of linear coefficients of length p, relating the outcome
Y to X.}

\item{effect_size}{The treatment effect, which is constant (not
heterogeneous).}
}
\value{
A list containing treatments A and outcomes Y.
}
\description{
Propensity scores pi are a logistic function of the covariates X. The
treatment is binary, 0 or 1. The outcomes are linear in the covariates, plus
independent Gaussian noise, plus a constant treatment effect for those who
received treatment. Given this structure, the treatment and the potential
outcomes are independent conditional on X, but not marginally because X
confounds the treatment and the outcome.
}
