% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_analysis.R
\name{simulate_inference}
\alias{simulate_inference}
\title{Compare performance of ATE estimators via simulation.}
\usage{
simulate_inference(counter, N, beta_a, beta_y, effect_size,
  estimators = c("aipw", "gcomp", "ipw", "tmle"), nonpar = FALSE,
  sl_lib_mu = NULL, sl_lib_pi = NULL, cv_folds = 5,
  bootstrap_reg = FALSE, bootNum = 100, oracle_estimators = NULL,
  remainder = FALSE)
}
\arguments{
\item{counter}{Integer used to set up randomization, via set.seed(), and to
name output files.}

\item{N}{The number of samples to generate.}

\item{beta_a}{A length 5 vector of coefficients, where the first coefficient
is the intercept.}
}
\description{
Compare performance of ATE estimators via simulation.
}
